+++ components/Carousel/CarouselImage/CarouselImage.tsx
