+++ components/Carousel/CarouselImage/index.ts
