+++ components/Carousel/index.ts
@@ -1,0 +1,1 @@
+export { Carousel } from './Carousel';
